point folks at hackathon branch of github repo
We're reserving the `hackathon-2020-04` branch for the hackathon: we'll
refrain from making any changes to that branch which might disrupt
participation in the hackathon. It will be more stable than the usual
`master` trunk, but will not get the same set of fixes.

We expect to delete the hackathon branch once the event is complete.